requirements.txt
Django==3.0.2
psycopg2==2.8.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
markdown==3.1.1
django-filter==2.2.0
django-polymorphic==2.1.2
youtube_dl==2020.1.15
